Adding document link test and removing the mocked features' tests
