--- testFixture/diagnostics.sh
@@ -1,1 +1,0 @@
-ASD asd
