+++ testFixture/include_test.sh
@@ -1,0 +1,1 @@
+include "test.sh"
